Update First.py Added confrimElement() method to TaskSet
confrimElement() checks if a given element is on the screen. This works however it currently searches the whole screen for the element and so it is a bit slow, but appears atleast accurate. The loop is inside the method thus the control mechanism does not need to reside in the main()

    def confirmElement(self,image):

        loop = True
        while loop:

            if ag.locateOnScreen(image) == None:
                continue
            else:
                loop = False

        return True