Updated DF website address
If only dayforce.com is used it goes to a newer designed web form which because of the new layout throws off the webpage loaded validator.  However by simply coding in the entire proper web address, the browser will fetch the originally designed Df log in page, and the validator will continue to work as always.